requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.15.0
tensorflow-hub>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0